C Project 4: Input Files and Calculating Points
In this project, I read input from an input file that has two initial points which are used to find the slope and Y-intercept. Then, the file has many X-values and the program must use the slope and Y-intercept to find the Y value that is parallel to the initial line. The equation used for this is Y=MX+B.