C Project 5: Structs, Input Files, Calculating Points
In this project, The end goal is the same as project 4 but the difference is that I used structs to get there. I read input from an input file that has two initial points which are used to find the slope and Y-intercept. Then, the file has many X-values and the program must use the slope and Y-intercept to find the Y value that is parallel to the initial line. The equation used to find the Y is Y=MX+B.