C Project 6: Deck
This program is designed to read through 2 decks of cards and return the longest amount of cards in ascending order (it can loop around going from K, A, 2, 3, 4, this is valid and is an ascending order of 5) and return the largest streak of cards with the same suit. The first line in the input represents the number of test cases (The most test cases there can be are 25). A test case consists of two lines (each 52 characters) with a number(2-A) followed by a suit (C, D, H, S).